requirements.txt
asgiref==3.5.2
dj-database-url==0.5.0
Django==4.0.4
django-cors-headers==3.12.0
django-heroku==0.3.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
engineering-notation==0.6.0
gunicorn==20.1.0
psycopg2==2.9.3
PyJWT==2.4.0
pytz==2022.1
sqlparse==0.4.2
tk-tools==0.15.0
tzdata==2022.1
whitenoise==6.1.0
